power: smbcharger-mmi: Fix OTG Enable
Phone fails to detect the Charger because the OTG
regulator is enabled and it can not properly sense
a change on vbus.

Make sure the the OTG Over-current IRQ only runs if
there is an active request Enabling OTG.  Also ensure
proper handling of the pulse skip state.  As a last
resort if the OTG is enabled when Charging is being
enabled be sure to Disable OTG.

Change-Id: I4561aa91869417fafc1a30753707b2b1d0866faa
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: https://gerrit.mot.com/1011959
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Ravikumar Vembu <raviv@motorola.com>
Submit-Approved: Jira Key